broke out basic math functions from std::math
The basic class TaskMath has templated functions for simple stuff that
will likely be reused. Since the class is bare except for a few public
members, it should be possible to inherit from it without conflicts
(unless you are using the cmath functions internally).

TODO: add a constructor and addOperation statements, so that child
classes can just call it without bloating the code with redundant
operations.